Some small code changes. Not processing all the feedback yet, as converting the project into Spring will require such changes anyway.

+++ SectionSpeedControl/src/main/java/org/exercise1/DataLine.java
@@ -20,9 +20,9 @@ public class DataLine {
         exitTimeMs = exitMs + 1000 * (exitSecond + 60 * (exitMinute + 60 * exitHour));
     }
 
-    public int getSpeed(double distance) {
+    public double getSpeed(double distance) {
         int elapsedMs = exitTimeMs - entranceTimeMs;
-        return  (int) (distance * 3600000 / elapsedMs);
+        return distance * 3600000 / elapsedMs;
     }
 
 }
